+++ api.go
@@ -1,38 +1,19 @@
 package main
 
-import (
-	"encoding/json"
-	"github.com/sirupsen/logrus"
-	"io"
-	"net/http"
-)
+import "fmt"
 
-func Get(uri string) (res map[string]interface{}, err error) {
-	req, err := http.NewRequest("GET", "https://api.github.com/"+uri, nil)
-	if err != nil {
-		return nil, err
-	}
+func GetUser(username string) (res map[string]interface{}, err error) {
+	return Get(fmt.Sprintf("users/%s", username))
+}
 
-	req.Header.Set("Accept", "application/json")
-	req.Header.Set("Authorization", "Bearer "+GetToken())
+func GetRepos(username string) (res map[string]interface{}, err error) {
+	return Get(fmt.Sprintf("users/%s/repos", username))
+}
 
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, err
-	}
+func GetRepo(username string, repo string) (res map[string]interface{}, err error) {
+	return Get(fmt.Sprintf("repos/%s/%s", username, repo))
+}
 
-	defer func(Body io.ReadCloser) {
-		err := Body.Close()
-		if err != nil {
-			logrus.Infoln()
-		}
-	}(resp.Body)
-
-	var data map[string]interface{}
-	err = json.NewDecoder(resp.Body).Decode(&data)
-	if err != nil {
-		return nil, err
-	}
-	return data, nil
+func GetLanguages(username string, repo string) (res map[string]interface{}, err error) {
+	return Get(fmt.Sprintf("repos/%s/%s/languages", username, repo))
 }
