+++ main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"fmt"
 	"github.com/sirupsen/logrus"
 )
 
@@ -10,4 +11,5 @@ func main() {
 
 	tokenList = GetTokenFromEnv()
 	ValidateToken()
+	fmt.Println(GetRepos("zmh-program"))
 }
