+++ utils.go
@@ -1,6 +1,12 @@
 package main
 
-import "fmt"
+import (
+	"encoding/json"
+	"fmt"
+	"github.com/sirupsen/logrus"
+	"io"
+	"net/http"
+)
 
 func Sum(array []int) int {
 	total := 0
@@ -39,4 +45,34 @@ func ScaleConvert(n float64, useSmallScale bool) string {
 		return fmt.Sprintf("%.0f", n)
 	}
 	return fmt.Sprintf("%.1f%s", n, scaleUnits[idx])
+}
+
+func Get(uri string) (res map[string]interface{}, err error) {
+	req, err := http.NewRequest("GET", "https://api.github.com/"+uri, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	req.Header.Set("Accept", "application/json")
+	req.Header.Set("Authorization", "Bearer "+GetToken())
+
+	client := &http.Client{}
+	resp, err := client.Do(req)
+	if err != nil {
+		return nil, err
+	}
+
+	defer func(Body io.ReadCloser) {
+		err := Body.Close()
+		if err != nil {
+			logrus.Infoln()
+		}
+	}(resp.Body)
+
+	var data map[string]interface{}
+	err = json.NewDecoder(resp.Body).Decode(&data)
+	if err != nil {
+		return nil, err
+	}
+	return data, nil
 }
